requirements.txt
google-generativeai
python-docxtpl
python-dotenv
